Add first SDK Create intent interaction
Just as a place to start. This isn't as tested as I would want a real
thing because it's just one big integration test instead of verifying
the client in a more thorough manner. Mostly because there is *no*
error handling proper.

But just to make more progress I won't be as strict about the quality.